å¹

+++ src/main/java/ball/ingram/demo/dto/QuestionDTO.java
@@ -5,7 +5,7 @@ import lombok.Data;
 
 @Data
 public class QuestionDTO {
-    private Integer id;
+    private long id;
     private String title;
     private String description;
     private long gmtCreate;
@@ -13,7 +13,7 @@ public class QuestionDTO {
     private Integer commentCount;
     private Integer viewCount;
     private Integer likeCount;
-    private int creator;
+    private long creator;
     private String tag;
     private User user;
 }
